+++ src/presenters/category.js
@@ -1,0 +1,6 @@
+module.exports = function ({ presenters, Presenter }) {
+  class CategoryPresenter extends Presenter {}
+  CategoryPresenter.type = 'category'
+  CategoryPresenter.plural = 'categories'
+  return CategoryPresenter
+}