+++ src/presenters/event.js
@@ -1,5 +1,6 @@
 module.exports = function ({ presenters, Presenter }) {
   class EventPresenter extends Presenter {}
-  EventPresenter.prototype.type = 'event'
+  EventPresenter.type = 'event'
+  EventPresenter.plural = 'events'
   return EventPresenter
 }