+++ src/presenters/ticket-batch.js
@@ -1,0 +1,6 @@
+module.exports = function ({ presenters, Presenter }) {
+  class TicketBatchPresenter extends Presenter {}
+  TicketBatchPresenter.type = 'ticketBatch'
+  TicketBatchPresenter.plural = 'ticketBatches'
+  return TicketBatchPresenter
+}