+++ src/presenters/workspace.js
@@ -1,5 +1,6 @@
 module.exports = function ({ presenters, Presenter }) {
   class WorkspacePresenter extends Presenter {}
-  WorkspacePresenter.prototype.type = 'workspace'
+  WorkspacePresenter.type = 'workspace'
+  WorkspacePresenter.plural = 'workspaces'
   return WorkspacePresenter
 }